shorten all file names
