+++ puck.files
@@ -1,43 +1,32 @@
 COPYING
+COPYING
 README.md
+README.md
+puck.cflags
+puck.cxxflags
 velodyne_puck/CMakeLists.txt
 velodyne_puck/package.xml
 velodyne_puck_decoder/CMakeLists.txt
-velodyne_puck_decoder/include/velodyne_puck_decoder/constants.h
-velodyne_puck_decoder/include/velodyne_puck_decoder/velodyne_puck_decoder.h
-velodyne_puck_decoder/include/velodyne_puck_decoder/velodyne_puck_decoder_nodelet.h
-velodyne_puck_decoder/launch/velodyne_puck_decoder_nodelet.launch
+velodyne_puck_decoder/launch/decoder_nodelet.launch
+velodyne_puck_decoder/launch/run.launch
 velodyne_puck_decoder/nodelet_velodyne_puck_decoder.xml
 velodyne_puck_decoder/package.xml
 velodyne_puck_decoder/src/constants.h
-velodyne_puck_decoder/src/velodyne_puck_decoder.cpp
-velodyne_puck_decoder/src/velodyne_puck_decoder.cpp
-velodyne_puck_decoder/src/velodyne_puck_decoder.cpp
-velodyne_puck_decoder/src/velodyne_puck_decoder.h
-velodyne_puck_decoder/src/velodyne_puck_decoder.h
-velodyne_puck_decoder/src/velodyne_puck_decoder_node.cpp
-velodyne_puck_decoder/src/velodyne_puck_decoder_node.cpp
-velodyne_puck_decoder/src/velodyne_puck_decoder_node.cpp
-velodyne_puck_decoder/src/velodyne_puck_decoder_nodelet.cpp
-velodyne_puck_decoder/src/velodyne_puck_decoder_nodelet.cpp
-velodyne_puck_decoder/src/velodyne_puck_decoder_nodelet.cpp
-velodyne_puck_decoder/src/velodyne_puck_decoder_nodelet.h
-velodyne_puck_decoder/src/velodyne_puck_decoder_nodelet.h
+velodyne_puck_decoder/src/decoder.cpp
+velodyne_puck_decoder/src/decoder.h
+velodyne_puck_decoder/src/decoder_node.cpp
+velodyne_puck_decoder/src/decoder_nodelet.cpp
+velodyne_puck_decoder/src/decoder_nodelet.h
 velodyne_puck_driver/CMakeLists.txt
-velodyne_puck_driver/include/velodyne_puck_driver/velodyne_puck_driver.h
-velodyne_puck_driver/include/velodyne_puck_driver/velodyne_puck_driver_nodelet.h
-velodyne_puck_driver/launch/velodyne_puck_driver_nodelet.launch
+velodyne_puck_driver/launch/driver_nodelet.launch
 velodyne_puck_driver/nodelet_velodyne_puck.xml
 velodyne_puck_driver/package.xml
 velodyne_puck_driver/src/constants.h
-velodyne_puck_driver/src/velodyne_puck_driver.cc
-velodyne_puck_driver/src/velodyne_puck_driver.cc
-velodyne_puck_driver/src/velodyne_puck_driver.h
-velodyne_puck_driver/src/velodyne_puck_driver_node.cc
-velodyne_puck_driver/src/velodyne_puck_driver_node.cc
-velodyne_puck_driver/src/velodyne_puck_driver_nodelet.cc
-velodyne_puck_driver/src/velodyne_puck_driver_nodelet.cc
-velodyne_puck_driver/src/velodyne_puck_driver_nodelet.h
+velodyne_puck_driver/src/driver.cpp
+velodyne_puck_driver/src/driver.h
+velodyne_puck_driver/src/driver_node.cpp
+velodyne_puck_driver/src/driver_nodelet.cpp
+velodyne_puck_driver/src/driver_nodelet.h
 velodyne_puck_msgs/CHANGELOG.rst
 velodyne_puck_msgs/CMakeLists.txt
 velodyne_puck_msgs/msg/VelodynePacket.msg
