+++ puck.includes
@@ -7,3 +7,11 @@ velodyne_puck_driver/include
 velodyne_puck_decoder/include/velodyne_puck_decoder
 velodyne_puck_driver/src
 velodyne_puck_decoder/src
+velodyne_puck
+velodyne_puck_msgs/msg
+velodyne_puck_driver/launch
+.
+velodyne_puck_msgs
+velodyne_puck_driver
+velodyne_puck_decoder/launch
+velodyne_puck_decoder
