+++ velodyne_puck/package.xml
@@ -1,7 +1,7 @@
 <package format="2">
 
   <name>velodyne_puck</name>
-  <version>2.0.0</version>
+  <version>2.0.1</version>
   <description>
     Basic ROS support for the Velodyne 3D LIDARs.
   </description>
