+++ velodyne_puck_decoder/package.xml
@@ -2,7 +2,7 @@
 <package format="2">
 
   <name>velodyne_puck_decoder</name>
-  <version>2.0.0</version>
+  <version>2.0.1</version>
   <description>
     Point cloud conversions for Velodyne 3D LIDARs.
   </description>
