+++ velodyne_puck_driver/CMakeLists.txt
@@ -12,14 +12,14 @@ find_package(catkin REQUIRED
 catkin_package()
 
 # Velodyne Puck driver
-add_library(${PROJECT_NAME} src/${PROJECT_NAME}.cc)
+add_library(${PROJECT_NAME} src/driver.cpp)
 target_include_directories(${PROJECT_NAME} PUBLIC ${catkin_INCLUDE_DIRS} src)
 target_link_libraries(${PROJECT_NAME} PUBLIC ${catkin_LIBRARIES})
 add_dependencies(${PROJECT_NAME} ${${PROJECT_NAME}_EXPORTED_TARGETS}
                  ${catkin_EXPORTED_TARGETS})
 
 # Velodyne Puck nodelet
-add_library(${PROJECT_NAME}_nodelet src/${PROJECT_NAME}_nodelet.cc)
+add_library(${PROJECT_NAME}_nodelet src/driver_nodelet.cpp)
 target_link_libraries(${PROJECT_NAME}_nodelet PUBLIC ${PROJECT_NAME})
 add_dependencies(${PROJECT_NAME}_nodelet ${${PROJECT_NAME}_EXPORTED_TARGETS}
                  ${catkin_EXPORTED_TARGETS})
