+++ velodyne_puck_driver/package.xml
@@ -1,7 +1,7 @@
 <package format="2">
 
   <name>velodyne_puck_driver</name>
-  <version>2.0.0</version>
+  <version>2.0.1</version>
   <description>
     ROS device driver for Velodyne PUCK VLP-16.
   </description>
