+++ velodyne_puck_driver/src/driver_nodelet.h
@@ -15,22 +15,27 @@
  * along with the driver.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <nodelet/nodelet.h>
 #include <ros/ros.h>
-#include "velodyne_puck_decoder.h"
 
-int main(int argc, char** argv) {
-  using namespace velodyne_puck_decoder;
-  ros::init(argc, argv, "velodyne_puck_decoder_node");
-  ros::NodeHandle nh;
-  ros::NodeHandle pnh("~");
+#include "driver.h"
 
-  VelodynePuckDecoder decoder(nh, pnh);
+namespace velodyne_puck_driver {
 
-  if (!decoder.initialize()) {
-    ROS_INFO("Cannot initialize the decoder...");
-    return -1;
-  }
+class VelodynePuckDriverNodelet : public nodelet::Nodelet {
+ public:
+  VelodynePuckDriverNodelet();
+  ~VelodynePuckDriverNodelet();
 
-  ros::spin();
-  return 0;
-}
+ private:
+  virtual void onInit();
+  virtual void devicePoll();
+
+  volatile bool running;  ///< device thread is running
+  boost::shared_ptr<boost::thread> device_thread;
+
+  VelodynePuckDriver::Ptr
+      velodyne_puck_driver;  ///< driver implementation class
+};
+
+}  // namespace velodyne_puck_driver
