+++ velodyne_puck_msgs/package.xml
@@ -1,7 +1,7 @@
 <package format="2">
 
   <name>velodyne_puck_msgs</name>
-  <version>2.0.0</version>
+  <version>2.0.1</version>
   <description>
     ROS message definitions for Velodyne 3D LIDARs.
   </description>
